% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cts}
\alias{cts}
\title{Estimated Salmon counts (6 vs 6)}
\format{matrix}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
cts
}
\description{
The estimated Salmon counts (scaledTPM) generated as shown in the
early steps in the workflow.
}
\keyword{datasets}
