% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dge.genes}
\alias{dge.genes}
\title{Genes with differential gene expression}
\format{character}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
dge.genes
}
\description{
The names of the genes in which the expression of all transcripts
was mulitplied by a common fold change.
}
\keyword{datasets}
