% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disps}
\alias{disps}
\title{Transcript-level dispersion parameter}
\format{numeric}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
disps
}
\description{
The dispersion parameter used to generate the transcript-level
Negative Binomial counts. Drawn from estimated dispersion from
GEUVADIS project samples. Matches the transcripts in \code{sim_counts}.
}
\keyword{datasets}
