% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dte.genes}
\alias{dte.genes}
\title{Genes with differential transcript expression}
\format{character}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
dte.genes
}
\description{
The names of the genes in which the expression of a single transcript
was multiplied by a fold change.
}
\keyword{datasets}
