% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtu.genes}
\alias{dtu.genes}
\title{Genes with differential transcript usage}
\format{character}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
dtu.genes
}
\description{
The names of the genes in which two transcripts had their
abundance (TPM) swapped. Two expressed transcripts were chosen if
possible, or one expressed and one non-expressed.
}
\keyword{datasets}
