% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dxr.g}
\alias{dxr.g}
\title{DEXSeq gene-level results}
\format{data.frame}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
dxr.g
}
\description{
DEXSeq gene-level results - see workflow for details.
}
\keyword{datasets}
