% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fold_changes}
\alias{fold_changes}
\title{Simulation fold changes}
\format{matrix}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
fold_changes
}
\description{
A matrix of fold changes for group 1 and group 2, as in the
format required for polyester input. The majority of the
fold changes are 1. Matches the transcripts in \code{sim_counts}.
}
\keyword{datasets}
