% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frag_GC_bias}
\alias{frag_GC_bias}
\title{Fragment GC bias}
\format{matrix}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
frag_GC_bias
}
\description{
The GC bias matrix as returned by plotGC in the alpine package. The columns
correspond to GEUVADIS samples as indicated in this workflow publication,
and each row corresponds to a percentile of GC, with 101 rows in total.
The first 12 columns were used to generate realistic fragment GC bias
using alpine and polyester.
}
\keyword{datasets}
