% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso.dge}
\alias{iso.dge}
\title{Isoforms in the genes with DGE}
\format{logical}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
iso.dge
}
\description{
A logical vector along all transcripts indicating which transcirpts
are in the genes with differential gene expression.
}
\keyword{datasets}
