% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso.dte.only}
\alias{iso.dte.only}
\title{Isoforms with DTE and not DTU}
\format{logical}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
iso.dte.only
}
\description{
A logical vector along all transcripts indicating which transcripts
are chosen for differential transcript expression, and which
are the only expressed transcript in the gene, thus not participating
in differential transcript usage.
}
\keyword{datasets}
