% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso.dte}
\alias{iso.dte}
\title{Isoforms with DTE}
\format{logical}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
iso.dte
}
\description{
A logical vector along all transcripts indicating which transcripts
are chosen for differential transcript expression.
}
\keyword{datasets}
