% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso.dtu}
\alias{iso.dtu}
\title{Isoforms with DTU}
\format{logical}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
iso.dtu
}
\description{
A logical vector along all transcripts indicating which transcripts
in the genes chosen for differential transcript usage are chosen
to have their abundance (TPM) swapped.
}
\keyword{datasets}
