% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim.counts.mat}
\alias{sim.counts.mat}
\title{Simulated counts}
\format{matrix}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
sim.counts.mat
}
\description{
The simulated counts for the two groups (columns).
Matches the transcripts in \code{sim_counts}.
}
\keyword{datasets}
