% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_counts}
\alias{sim_counts}
\title{Simulated counts for group 1}
\format{numeric}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
sim_counts
}
\description{
The simulated counts as provided to polyester (representing
the base line counts for group 1). The represented transcripts
are a subset of all transcripts in the reference, as we do not
simulate from transcripts with an expected count of 5 or more.
}
\keyword{datasets}
