% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tpms}
\alias{tpms}
\title{Simulated abundances}
\format{matrix}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
tpms
}
\description{
The simulated abundances (TPM) for the two groups (columns).
The transcripts correspond to the human transcripts from
GENCODE version 28.
}
\keyword{datasets}
