% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txdf}
\alias{txdf}
\title{Transcript-to-gene data.frame}
\format{data.frame}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
txdf
}
\description{
A table matching transcript to gene, and providing the number
of transcripts per gene. The GTF for human from GENCODE version 28
was used to generate this data.frame.
}
\keyword{datasets}
