% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txi.g}
\alias{txi.g}
\title{Gene-level tximport object}
\format{list}
\source{
\url{http://dx.doi.org/10.12688/f1000research.15398.1}
}
\usage{
txi.g
}
\description{
The tximport object, summarized to gene level, generated
as shown in the DGE section of the workflow.
}
\keyword{datasets}
